feat: implement Feed and SubCategory screens with automated highlight carousel and category navigation utility
